Re-organize the component directory
