+++ src/lib/components/cards/MiniCardTicks.svelte
