+++ src/lib/components/cards/RankInfoBox.svelte
